+++ modules/aws/database/rds/postgres/data.tf
@@ -1,3 +1,1 @@
 data "aws_partition" "current" {}
-
-data "aws_region" "current" {}